requirements.txt
streamlit
firebase-admin
google-generativeai
bcrypt
extra-streamlit-components
xhtml2pdf
markdown
pandas
streamlit-calendar
toml